@@ -1,18 +1,16 @@
-﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 
 namespace Calculator
 {
     public static class Program
     {
-        public static void Main(string[] args)
-        {
+        public static void Main(string[] args) =>
             BuildWebHost(args).Run();
-        }
 
-        private static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
+        private static IHost BuildWebHost(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(web => web.UseStartup<Startup>())
                 .Build();
     }
 }
